Simplified the signal handling whenever a node's status changes.
In the template code you were observing an event 'node' on which
you *directly* did a postback with a certain action to do the animations.

That felt unnecessary: the "take_down" and "bring_up" actions are not
really needed unless you want to do animation.

To simplify, I just refreshed the _left and _right templates from the
page callback and that's it.  I renamed the emitted event to
'node_status_changed'.

However, your approach is correct if you want to be able to do these
types of animation.